refactor: split workboard persistence types
